if email change request retry delay not expired, display error...
message instead of misleading success message (misleading because it always displayed data currently in user.emailChangePending without updating it even if user entered another email address during his successive attempts while retry delay is still not expired.
Updating the data is not an option because it could potentially lead to a verification link sent to an address but clicked while another address is in user.emailChangePending.